backend/requirements.txt
Flask>=3.1.1
Flask-JWT-Extended>=4.5.2
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
PyMySQL>=1.0.3
cryptography>=41.0.0
Werkzeug>=2.3.0
flask-socketio>=5.5.1
google-generativeai>=0.8.5
python-dotenv>=1.1.1
 flask_cors>=6.0.1